feat: IDP-1106 findById optimize to be run on cosmosDB (#68)
* IDP-1106 findById optimize to be run on cosmosDB

* extending coverage
